getParam quick method added, RequestResponse class package changed, contentType added to response, filter and register methods added to Server to filter requests and add handlers for http response statuses, a swagger.yaml example file added, readme updated

+++ src/test/java/fatjar/test/WebTester.java
@@ -2,9 +2,11 @@ package fatjar.test;
 
 import fatjar.*;
 import fatjar.dto.HttpMethod;
+import fatjar.dto.Status;
 import sample.MyEntity;
 
 import java.util.List;
+import java.util.Optional;
 
 public class WebTester {
 
@@ -17,6 +19,17 @@ public class WebTester {
         // do a request to http://localhost:8080/ and/or http://localhost:8080/Hi
         Server.create()
                 .listen(8080, "localhost")
+                .register(Status.STATUS_BAD_REQUEST, (req, res) -> {
+                    res.setContentType("text/html");
+                    res.setContent("<h1>BAD REQUEST!</h1>");
+                    res.write();
+                })
+                .filter("/*", (req, res) -> {
+                    Log.info("Wildcard filter called");
+                })
+                .filter("/Hi", (req, res) -> {
+                    Log.info("/Hi filter called");
+                })
                 .get("/", (req, res) -> {
                     res.setContent("Welcome");
                     res.write();
@@ -45,17 +58,21 @@ public class WebTester {
                     Log.info("fount entityID1L: " + entityID1L);
 
                     MyEntity entity = new MyEntity("john");
-                    MyEntity insertedEntity = db.insert(entity);
-                    Log.info("insert insertedEntity: " + insertedEntity);
+                    Optional<MyEntity> insertedEntityOptional = db.insert(entity);
+                    if (insertedEntityOptional.isPresent()) {
+                        MyEntity insertedEntity = insertedEntityOptional.get();
+                        Log.info("insert insertedEntity: " + insertedEntity);
 
-                    insertedEntity.setName("johnny");
-                    db.update(insertedEntity);
-                    Log.info("update insertedEntity: " + insertedEntity);
+                        insertedEntity.setName("johnny");
+                        db.update(insertedEntity);
+                        Log.info("update insertedEntity: " + insertedEntity);
 
-                    db.delete(insertedEntity);
-                    Log.info("delete insertedEntity: " + insertedEntity);
-
-                    res.setContent(JSON.toJson(insertedEntity));
+                        db.delete(insertedEntity);
+                        Log.info("delete insertedEntity: " + insertedEntity);
+                        res.setContent(JSON.toJson(insertedEntity));
+                    } else {
+                        res.setContent("could not insert to DB");
+                    }
                     res.write();
                 })
                 .get("/httpClient", (req, res) -> {
@@ -71,8 +88,14 @@ public class WebTester {
                     }
                     res.write();
                 })
+                .get("/throwException", (req, res) -> {
+                    throw new Server.ServerException("tojsonexception");
+                })
+                .get("/badRequest", (req, res) -> {
+                    throw new Server.ServerException(Status.STATUS_BAD_REQUEST);
+                })
                 .get("/toJSON", (req, res) -> {
-                    res.setContent(JSON.toJson(new MyPOJO("john", 101)));
+                    res.setContent(JSON.toJson(new MyPOJO(req.getParam("name", "add a query parameter like '?name=john'"), 101)));
                     res.write();
                 })
                 .post("/fromJSON", (req, res) -> {
